Move '{R,r}easonblySized' to Prelude (from Test.Prelude) to use it for generators in source code.
