+++ hydra-test-utils/src/Test/Hydra/Prelude.hs
@@ -4,11 +4,9 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 
 module Test.Hydra.Prelude (
@@ -47,7 +45,7 @@ import Test.HSpec.JUnit (junitFormat)
 import Test.HUnit.Lang (FailureReason (Reason), HUnitFailure (HUnitFailure))
 import Test.Hspec.Core.Format (Format, FormatConfig (..))
 import Test.Hspec.Core.Formatters (formatterToFormat, specdoc)
-import Test.QuickCheck (Property, Testable, coverTable, forAll, scale, tabulate)
+import Test.QuickCheck (Property, Testable, coverTable, forAll, tabulate)
 
 -- | Create a unique temporary directory.
 createSystemTempDirectory :: String -> IO FilePath
@@ -125,33 +123,6 @@ checkProcessHasNotDied name processHandle =
   waitForProcess processHandle >>= \case
     ExitSuccess -> pure ()
     ExitFailure exit -> failure $ "Process " <> show name <> " exited with failure code: " <> show exit
-
--- | Resize a generator to grow with the size parameter, but remains reasonably
--- sized. That is handy when testing on data-structures that can be arbitrarily
--- large and, when large entities don't really bring any value to the test
--- itself.
---
--- It uses a square root function which makes the size parameter grows
--- quadratically slower than normal. That is,
---
---     +-------------+------------------+
---     | Normal Size | Reasonable Size  |
---     | ----------- + ---------------- +
---     | 0           | 0                |
---     | 1           | 1                |
---     | 10          | 3                |
---     | 100         | 10               |
---     | 1000        | 31               |
---     +-------------+------------------+
-reasonablySized :: Gen a -> Gen a
-reasonablySized = scale (ceiling . sqrt @Double . fromIntegral)
-
--- | A QuickCheck modifier to make use of `reasonablySized` on existing types.
-newtype ReasonablySized a = ReasonablySized a
-  deriving newtype (Show, ToJSON, FromJSON)
-
-instance Arbitrary a => Arbitrary (ReasonablySized a) where
-  arbitrary = ReasonablySized <$> reasonablySized arbitrary
 
 -- | Like 'coverTable', but construct the weight requirements generically from
 -- the provided label.
